Improve submdspan testing (#342)
This now tests that the elements of the submdspan point to the correct
element of the src. That caught a mistake in the layout_foo test layout.
Disable bracket operator for icpx due to compiler crash just in the test config.
Note: for icpx 2024 we could turn it on again.